Fix potential bug in player target_next
Incrementing the target_index after selecting the next target could cause an
error if the index is one less than the targets container size, and one of
the targets is destroyed before the player changes targets: the index would
be invalid for the size of the collection. Instead, we increment the target
index when selecting the next target, but only if we already have a target
selected.